+++ packages/ai/core/generate-text/generate-text.ts
@@ -1,6 +1,6 @@
 import { createIdGenerator } from '@ai-sdk/provider-utils';
 import { Tracer } from '@opentelemetry/api';
-import { InvalidArgumentError } from '../../errors';
+import { InvalidArgumentError, ToolExecutionError } from '../../errors';
 import { CoreAssistantMessage, CoreMessage, CoreToolMessage } from '../prompt';
 import { CallSettings } from '../prompt/call-settings';
 import { convertToLanguageModelPrompt } from '../prompt/convert-to-language-model-prompt';
@@ -29,6 +29,7 @@ import { StepResult } from './step-result';
 import { toResponseMessages } from './to-response-messages';
 import { ToolCallArray } from './tool-call';
 import { ToolResultArray } from './tool-result';
+import { ToolCallRepairFunction } from './tool-call-repair';
 
 const originalGenerateId = createIdGenerator({ prefix: 'aitxt', size: 24 });
 
@@ -97,6 +98,7 @@ export async function generateText<
   experimental_telemetry: telemetry,
   experimental_providerMetadata: providerMetadata,
   experimental_activeTools: activeTools,
+  experimental_repairToolCall: repairToolCall,
   _internal: {
     generateId = originalGenerateId,
     currentDate = () => new Date(),
@@ -155,6 +157,11 @@ changing the tool call and result types in the result.
     experimental_activeTools?: Array<keyof TOOLS>;
 
     experimental_output?: Output<OUTPUT>;
+
+    /**
+A function that attempts to repair a tool call that failed to parse.
+     */
+    experimental_repairToolCall?: ToolCallRepairFunction<TOOLS>;
 
     /**
     Callback that is called when each step (LLM call) is finished, including intermediate steps.
@@ -354,8 +361,16 @@ changing the tool call and result types in the result.
         );
 
         // parse tool calls:
-        currentToolCalls = (currentModelResponse.toolCalls ?? []).map(
-          modelToolCall => parseToolCall({ toolCall: modelToolCall, tools }),
+        currentToolCalls = await Promise.all(
+          (currentModelResponse.toolCalls ?? []).map(toolCall =>
+            parseToolCall({
+              toolCall,
+              tools,
+              repairToolCall,
+              system,
+              messages: stepInputMessages,
+            }),
+          ),
         );
 
         // execute tools:
@@ -555,31 +570,39 @@ async function executeTools<TOOLS extends Record<string, CoreTool>>({
         }),
         tracer,
         fn: async span => {
-          const result = await tool.execute!(args, {
-            toolCallId,
-            messages,
-            abortSignal,
-          });
-
           try {
-            span.setAttributes(
-              selectTelemetryAttributes({
-                telemetry,
-                attributes: {
-                  'ai.toolCall.result': {
-                    output: () => JSON.stringify(result),
-                  },
-                },
-              }),
-            );
-          } catch (ignored) {
-            // JSON stringify might fail if the result is not serializable,
-            // in which case we just ignore it. In the future we might want to
-            // add an optional serialize method to the tool interface and warn
-            // if the result is not serializable.
-          }
+            const result = await tool.execute!(args, {
+              toolCallId,
+              messages,
+              abortSignal,
+            });
 
-          return result;
+            try {
+              span.setAttributes(
+                selectTelemetryAttributes({
+                  telemetry,
+                  attributes: {
+                    'ai.toolCall.result': {
+                      output: () => JSON.stringify(result),
+                    },
+                  },
+                }),
+              );
+            } catch (ignored) {
+              // JSON stringify might fail if the result is not serializable,
+              // in which case we just ignore it. In the future we might want to
+              // add an optional serialize method to the tool interface and warn
+              // if the result is not serializable.
+            }
+
+            return result;
+          } catch (error) {
+            throw new ToolExecutionError({
+              toolName,
+              toolArgs: args,
+              cause: error,
+            });
+          }
         },
       });