+++ packages/ai/core/generate-text/run-tools-transformation.ts
@@ -1,6 +1,7 @@
 import { LanguageModelV1StreamPart } from '@ai-sdk/provider';
 import { generateId } from '@ai-sdk/ui-utils';
 import { Tracer } from '@opentelemetry/api';
+import { ToolExecutionError } from '../../errors';
 import { NoSuchToolError } from '../../errors/no-such-tool-error';
 import { CoreMessage } from '../prompt/message';
 import { assembleOperationName } from '../telemetry/assemble-operation-name';
@@ -17,6 +18,7 @@ import {
 import { calculateLanguageModelUsage } from '../types/usage';
 import { parseToolCall } from './parse-tool-call';
 import { ToolCallUnion } from './tool-call';
+import { ToolCallRepairFunction } from './tool-call-repair';
 import { ToolResultUnion } from './tool-result';
 
 export type SingleRequestTextStreamPart<
@@ -67,16 +69,20 @@ export function runToolsTransformation<TOOLS extends Record<string, CoreTool>>({
   toolCallStreaming,
   tracer,
   telemetry,
+  system,
   messages,
   abortSignal,
+  repairToolCall,
 }: {
   tools: TOOLS | undefined;
   generatorStream: ReadableStream<LanguageModelV1StreamPart>;
   toolCallStreaming: boolean;
   tracer: Tracer;
   telemetry: TelemetrySettings | undefined;
+  system: string | undefined;
   messages: CoreMessage[];
   abortSignal: AbortSignal | undefined;
+  repairToolCall: ToolCallRepairFunction<TOOLS> | undefined;
 }): ReadableStream<SingleRequestTextStreamPart<TOOLS>> {
   // tool results stream
   let toolResultsStreamController: ReadableStreamDefaultController<
@@ -120,7 +126,7 @@ export function runToolsTransformation<TOOLS extends Record<string, CoreTool>>({
     LanguageModelV1StreamPart,
     SingleRequestTextStreamPart<TOOLS>
   >({
-    transform(
+    async transform(
       chunk: LanguageModelV1StreamPart,
       controller: TransformStreamDefaultController<
         SingleRequestTextStreamPart<TOOLS>
@@ -187,9 +193,12 @@ export function runToolsTransformation<TOOLS extends Record<string, CoreTool>>({
           }
 
           try {
-            const toolCall = parseToolCall({
+            const toolCall = await parseToolCall({
               toolCall: chunk,
               tools,
+              repairToolCall,
+              system,
+              messages,
             });
 
             controller.enqueue(toolCall);
@@ -257,7 +266,11 @@ export function runToolsTransformation<TOOLS extends Record<string, CoreTool>>({
                     (error: any) => {
                       toolResultsStreamController!.enqueue({
                         type: 'error',
-                        error,
+                        error: new ToolExecutionError({
+                          toolName: toolCall.toolName,
+                          toolArgs: toolCall.args,
+                          cause: error,
+                        }),
                       });
 
                       outstandingToolResults.delete(toolExecutionId);