+++ packages/groq/src/groq-chat-settings.ts
@@ -1,15 +1,10 @@
 // https://console.groq.com/docs/models
+// production models
 export type GroqChatModelId =
   | 'gemma2-9b-it'
   | 'gemma-7b-it'
-  | 'llama3-groq-70b-8192-tool-use-preview'
-  | 'llama3-groq-8b-8192-tool-use-preview'
-  | 'llama-3.1-70b-versatile'
+  | 'llama-3.3-70b-versatile'
   | 'llama-3.1-8b-instant'
-  | 'llama-3.2-1b-preview'
-  | 'llama-3.2-3b-preview'
-  | 'llama-3.2-11b-vision-preview'
-  | 'llama-3.2-90b-vision-preview'
   | 'llama-guard-3-8b'
   | 'llama3-70b-8192'
   | 'llama3-8b-8192'