+++ packages/provider/src/index.ts
@@ -1,5 +1,6 @@
 export * from './embedding-model/index';
 export * from './errors/index';
+export * from './image-model/index';
 export * from './json-value/index';
 export * from './language-model/index';
 export * from './provider/index';