+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,34 @@
 
 ### Patch Changes
 
+- Updated dependencies [09a9cab]
+  - ai@4.0.14
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [9f32213]
+  - ai@4.0.13
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5167bec]
+- Updated dependencies [0984f0b]
+  - ai@4.0.12
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.0.11
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [913872d]
   - ai@4.0.10