+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ai-sdk/azure
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [09a9cab]
+  - @ai-sdk/provider@1.0.2
+  - @ai-sdk/openai@1.0.8
+  - @ai-sdk/provider-utils@2.0.4
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [0984f0b]
+  - @ai-sdk/provider-utils@2.0.3
+  - @ai-sdk/openai@1.0.7
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [a9a19cb]
+  - @ai-sdk/openai@1.0.6
+
 ## 1.0.7
 
 ### Patch Changes